move config to .mts extension
