+++ packages/cli/src/setup/detect-setup.mts
@@ -1,5 +1,5 @@
 import { resolve } from 'path'
-import { getConfigWithDefaultValues } from '../../../config/src/config'
+import { getConfigWithDefaultValues } from '../../../config/src/config.mjs'
 import type { Adapters, GeneratorConfig } from '../../../config/src/types.mjs'
 import { doesPathExist } from '../../../generator/src/utils/file.utils.mjs'
 import { getDependencyList, isEsmProject } from './package-json.mjs'
