+++ packages/cli/src/setup/setup.mts
@@ -2,7 +2,7 @@ import { diff as justDiff } from 'just-diff'
 import { diffApply as justDiffApply } from 'just-diff-apply'
 import kleur from 'kleur'
 import { isPropertyNotUndefined } from 'typesafe-utils'
-import { doesConfigFileExist, getConfigWithDefaultValues, writeConfigToFile } from '../../../config/src/config'
+import { doesConfigFileExist, getConfigWithDefaultValues, writeConfigToFile } from '../../../config/src/config.mjs'
 import type { GeneratorConfig } from '../../../config/src/types.mjs'
 import { logger } from '../../../generator/src/utils/logger.mjs'
 import { getDefaultConfig } from './detect-setup.mjs'
