+++ packages/config/package.json
@@ -8,5 +8,6 @@
     "ajv": "^6.12.6",
     "kleur": "^4.1.5",
     "typescript-json-schema": "^0.53.1"
-  }
+  },
+  "type": "module"
 }
