+++ packages/config/src/validation.mts
