+++ packages/generator/src/generator.ts
@@ -3,7 +3,7 @@ import { execSync } from 'child_process'
 import { watch } from 'chokidar'
 import { resolve } from 'path'
 import ts from 'typescript'
-import { getConfigWithDefaultValues, readConfig } from '../../config/src/config'
+import { getConfigWithDefaultValues, readConfig } from '../../config/src/config.mjs'
 import type { GeneratorConfig, GeneratorConfigWithDefaultValues } from '../../config/src/types.mjs'
 import { generate } from './generate-files'
 import { configureOutputHandler, shouldGenerateJsDoc } from './output-handler.mjs'
